+++ src/lib.rs
@@ -160,5 +160,10 @@ pub mod spi;
 pub mod time;
 #[cfg(feature = "device-selected")]
 pub mod timer;
+#[cfg(all(
+    feature = "stm32-usbd",
+    any(feature = "stm32f102", feature = "stm32f103")
+))]
+pub mod usb;
 #[cfg(feature = "device-selected")]
 pub mod watchdog;